update collateral WETH manager

+++ contracts/core/diaspore/utils/IWETH9.sol
@@ -5,5 +5,5 @@ import "../../../interfaces/IERC20.sol";
 
 contract IWETH9 is IERC20 {
     function deposit() public payable;
-    function withdraw(uint wad) public;
+    function withdraw(uint256 wad) public;
 }
